refactor: filter option
